+++ Arduino/sketches/report_temp/network.h
@@ -1,7 +1,8 @@
 #ifndef NETWORK_H
 #define NETWORK_H
+#include <stdint.h>
 #include "statemachine.h"
 #include "sharedstate.h"
 
-StateMachineChoice initializeNetwork(unsigned long sendEvery, SharedState* sharedState);
+StateMachineChoice initializeNetwork(uint32_t sendEvery, SharedState* sharedState);
 #endif
